spring security config oauth2 jwt

+++ src/main/java/com/vde/gestionetudiants/repository/EtudiantRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EtudiantRepository extends JpaRepository<Etudiant, Integer> {
+public interface EtudiantRepository extends JpaRepository<Etudiant, Long> {
 }
